🚧 Fix database issues with foreign keys and relationships, also add db tests (#117)
## Describe your changes

- Removed redundant foreign keys
- Named foreign keys for clarity
- Added new association table
- @aaronhaefner I modified the db tests by removing some columns that
were not necessary, I also modified the e-mail check to only check for
the "@" character.
- Updated poetry.lock with db test requirements.


## Non-obvious technical information

NOTE: A new bug was introduced for the migrations. I had to change this
line `from new_arrivals_chi.app.main import db
` in `database.py` to: `from main import db` to get the migrations to
work. However, if the migrations work, the test will fail,


## Checklist before requesting a review
- [x] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [x] Link issue
- [x] Update relevant documentation if applicable: doc strings, readme,
poetry.