User sign up functionality and injection into database (#119)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes
Closes #78
Allows for user to sign up and be added to the user data base
Validated inputs to ensure they match specifications

## Non-obvious technical information
Still questioning whether we should normalize inputs and if this level
of sanitization is enough. Creating a ticket for futher research.

Planning to add addition prs that branch off of this to make them
smaller.

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.

---------

Co-authored-by: Federico Domínguez Molina <fd.molina@outlook.com>
Co-authored-by: madeleineroberts <C@nney3>