Login_functionality (#120)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes
Closes #45, #72, #77 

Implements the login and log out functionality for a user.

Adds in login_required for pages that should not be able to be accessed
as an unauthorized user.

## Non-obvious technical information
Once a user is signed up, they are able to login to their account via
the login page. When logged in the user is able to access the profile
page. User can log out of their account by pressing the logout button;
the functionality of this is confirmed as the profile page now directs
the user to log in.

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.

---------

Co-authored-by: Federico Domínguez Molina <fd.molina@outlook.com>
Co-authored-by: madeleineroberts <C@nney3>