Implement change password logic, skeleton HTML, and update the DB reference (#122)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes
Closes #46

Implemented the ablity for a user to be able to update their password. 

To update thier password:
1) a user must be authenticated
2) user must enter old password
3) user must enter their new password that matches the password
contraints twice
On success, a hashed verison of the previous password replaces their
previous password in the database.

Once their new password is validated, the user can no longer log in with
their previous password.


## Non-obvious technical information
http://127.0.0.1:5000/change_password
Added button to access this route within the user authenicated profile,
rather than on the log in page as is suggested by the ticket. Will plan
to add a forgot password function to the log in page.

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.

---------

Co-authored-by: Federico Domínguez Molina <fd.molina@outlook.com>
Co-authored-by: madeleineroberts <mkroberts@uchicago.edu>