Add pytest fixtures for database and logger, use in db_tests.py (#130)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes

This PR implements additional fixtures that can be used for testing,
`database` and `logger` fixtures. There is also now a way to override
the app configuration to use a testing configuration instead upon
creation. The documentation is updated to reflect the new fixtures and
tests.

## Non-obvious technical information

## Checklist before requesting a review
- [x] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [ ] Link issue
- [x] Update relevant documentation if applicable: doc strings, readme,
poetry.