Ensure testing files are interacting with testing instance of database (#132)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes

This PR makes changes to testing files to ensure that the functions and
fixtures are interacting with a testing instance of the database.

The database was unexpectedly deleted, so we need to confirm that
running these tests has no impact on the production database.

## Non-obvious technical information

Running `pytest` currently produces the following **warning**, which is
not seen on the `main` branch. This does not affect the app, but we may
need to investigate with a new issue.
```
tests/db_test.py::test_user_organization_creation
  /Users/aaronhaefner/Library/Caches/pypoetry/virtualenvs/new-arrivals-chi-WfKZdJZx-py3.11/lib/python3.11/site-packages/flask_sqlalchemy/extension.py:881: 
SAWarning: Can't sort tables for DROP; an unresolvable foreign key dependency exists between tables: 
hours, locations, organizations, users; and backend does not support ALTER. 
To restore at least a partial sort, apply use_alter=True to ForeignKey and ForeignKeyConstraint objects involved 
in the cycle to mark these as known cycles that will be ignored.
    getattr(metadata, op_name)(bind=engine)
```

## Checklist before requesting a review
- [x] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [ ] Link issue
- [ ] Update relevant documentation if applicable: doc strings, readme,
poetry.

---------

Co-authored-by: Federico Domínguez Molina <67844597+FedericoDM@users.noreply.github.com>