(test) Authentication Tests (#153)
This pull request creates a series of tests for the authorization routes
and provides some security tests for authentication.

This PR will close #126 and close #127

## Describe your changes

Updated a few things in the conftest.py file:
- We need a secret key for the in-memory DB to run, added a placeholder
key for testing
- Added 3 fixtures:
- capture_templates: extract the template of the current page, so we can
assert that the proper template is rendered, after a successful 200
response
- test_user: creates an instance of the test user in the database. This
is needed to test the login functionality for a user already in the
database, as well as to test the password changes
- login_client: creates a logged-in state, to test functionalities that
require the user already be logged in (i.e. change password, etc.)

Three different test files: 
- auth_basic_test.py: tests to check login, logout, and sign-up
functionality
- auth_change_password_test.py: a suite of tests specifically focused on
different scenarios under which the password is changed
- auth_security_test.py: additional tests to put pressure on
authentication security

## Non-obvious technical information


## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.