Require HTTPS to implement secure HTTP requests (#154)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes
Transitions the development server from running our application using
HTTP protocol to HTTPS. This will guarantee that all HTTP requests and
responses are encrypted, enhancing the security of data transmitted over
the network.
closes #155 

## Non-obvious technical information
For the development server, we are using a auto-generated self-signed
certificate as a result the CA is unable to validate a server
certificate, though you can continue to proceed and visit the
development site. For the production deployment, we will ensure a valid
certificate from CA for our domain.

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.