(feat) functionality to add basic org info to db (#167)
This PR adds basic organization info to the database and implements
tests for this functionality. This PR will close #115

## Describe your changes
- create_organization_profile in data_handler.py will check that the
required info has been provided and if so, will create a new org in the
database, returning the new org ID
- /profile route has been updated in main.py to allow for both GET and
POST to be used. When logged in, /profile will default to GET where it
gets the org information for the org connected to the current user. When
post is called, the org info will be updated for the current user, and
then the profile page will be returned with the updated info
- org_test.py adds tests for this 

## Other changes not directly related to this issue:
- I made a slight change to have the capture_templates fixture is used
in all the tests I've written. Rather than indexing into the specific
template, I will now determine how many templates were captured and only
return the final one

```
 final_template_rendered = len(capture_templates) - 1
        assert (
            capture_templates[final_template_rendered][0].name == "signup.html"
        ), "Wrong template used"
```

- added the logger to the xxs_test.py file
- made a small fix to a typo in a docstring in the security testing file


## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.