Filterable Table with Data (#134)
## Describe your changes

This PR uses the fake database generated by setup_fake_data.py for
demonstrative purposes.

This PR builds upon the previously created health_search.html by
creating a filterable table using fake data generated with
setup_fake_data.py.

Furthermore, this PR addresses broadly the translation in-place issues
detailed by PR #100 by updating base.html to be dynamic to which page a
user is currently on. It removes manual coding of translation
redirection in-place (such as pressing "English" in the legal page takes
you to the legal page, but rendered in Spanish.)

The following files were primarily amended:
- main.css, to add additional styling to the table
- main.py, to integrate fake database call into route 
- health_search.html, to create filtering of table using passed data
from route
- base.html, to generalize translation logic
- numerous .html files to remove hard-coded translation logic

The following files & directories were created:
- static/js, to store JavaScript files
- filter.js, to perform filtering and hiding

Accepting this PR closes #28 and closes #100.

## Non-obvious technical information

The current fake database does not generate all information, such as
neighborhood and supplies. Comments were left in main.py to signify
where changes would need to be made. The following files would need to
be updated in order to incorporate changes with the real database:

- main.py, to adjust database call in route to the real database
- filter.js, to handle new parameters passed in from main.py route
method

Since we have a .js file now, it may be worthwhile to incorporate a
linter for .js files.

Future improvements need to either limit the number of rows initially
shown, or create scrolling within the table to avoid the table being too
large on smaller displays.

Also, translations need to be created in the .json. 

## Checklist before requesting a review
- [x] pre-commit run --all-files (run before pushing)
- [x] pytest if applicable
- [x] Link issue
- [x] Update relevant documentation if applicable: doc strings, readme,
poetry.