SQL injection correction (#203)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes

This request closes #128. I added sql injections to three functions:
- test_login_invalid_credentials
- test_signup_post_invalid_password
- test_signup_post_invalid_email

Also, i updated the README so that the sql_injection file bullet is
taken out and just kept it in the "tests included" section.

@michplunkett, I created a new post within each function because I
didn't want to take away the functionality of the original test. I added
a new post to each function to see if it would pass the original test
plus a sql injection to that login or signup.

Michael also added some constants and removed duplicated constants.


## Non-obvious technical information
n/a

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.

---------

Co-authored-by: Xiomara Salazar <xiomarasalazar@Xiomaras-Big-MacBook-Pro.local>
Co-authored-by: michplunkett <5885605+michplunkett@users.noreply.github.com>