(feat) legal static pages (#223)
This PR will primarily close #198 and #201 

It creates all the static pages for the end of the legal flows. The
Spanish translation is available but it is preliminary (translation was
done by ChatGPT to demonstrate functionality, will require more detailed
translation before deployment)

It will also close #222, a bug fix.

## Describe your changes
- Created loop in jinja to read in test with different keys for
formatting and format accordinly
  - Can hand title, header, subheader, bullet list, body text and links
- Updated the en and es jsons with all the legal text mapped with keys
to indicate formatting
- created the HTML for each of the pages with this loop
- created CSS to formatting the legal static pages
  
## Non-obvious technical information
- Also fixed an untracked bug that the "back" button of the legal flow
where the color wasn't being retained on return to a page
- Minor change to one of the legal tests to account for a change in file
paths

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.