Editing edit org template (#242)
<!---
Please write your PR name in the present imperative tense. Examples of
present imperative tense are:
"Fix issue in the dispatcher where…", "Improve our handling of…", etc."

For more information on Pull Requests, you can reference here:

https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe your changes
closes #212
Creates the HTML for the edit organization dashboard. 

## Non-obvious technical information
Will need to determine still if we want to allow the user to change
email -- this is currently not functional with extract org change.
Routes and buttons are currently not functional and will not change
information in the db.


## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.

```commandline
HERE IS SOME COMMAND LINE OUTPUT
```

---------

Co-authored-by: Summer Long <salong@uchicago.edu>
Co-authored-by: Summer Long <91840911+sumslong@users.noreply.github.com>
Co-authored-by: Federico Domínguez Molina <fd.molina@outlook.com>
Co-authored-by: klinkoberstar <111700741+klinkoberstar@users.noreply.github.com>