Translation Refactor with flask-babel (#283)
This PR removes the use of a series of translation JSONs in favor of
flask-babel for localization. This PR will close #262

## Describe your changes
- Sets up flask-babel in main.py
- Refactors HTML templates to use lazygettext to extract strings
- Updates the routes to remove explicit calling to translation JSONs
- Updates tests 
- Updates the README with instructions about how to extract strings,
update translations and compile translations
- Removes old translations, and all helper functions associated with the
old version of managing translations

## Non-obvious technical information
- Not all strings have translations - the app will default to English
text if a translation is not present. Every page should have at least
one piece of text that will be translated to demonstrate functionality
but not all english strings have a placeholder in Spanish at this time.
Official translations will be completed as a separate PR. This PR
focuses on functionality.
- The translations currently running in the app have not been verified
and likely contain inaccuracies.

## Checklist before requesting a review
- [X] pre-commit run --all-files (run before pushing)
- [X] pytest if applicable
- [X] Link issue
- [X] Update relevant documentation if applicable: doc strings, readme,
poetry.